check if holepunch and query are set as well before destroying them
re comment https://github.com/hyperswarm/dht/pull/47#issuecomment-908263844

+++ index.js
@@ -216,8 +216,8 @@ module.exports = class HyperDHT extends DHT {
 
     function ontimeout () {
       if (!error) error = TIMEOUT
-      query.destroy()
-      holepunch.destroy()
+      if (query) query.destroy()
+      if (holepunch) holepunch.destroy()
     }
   }
